requirements.txt
Django~=4.2.0
django-environ~=0.11.0
django-cors-headers~=4.4.0
djangorestframework~=3.15.0

argon2-cffi~=23.1.0
djoser~=2.2.0
djangorestframework-simplejwt~=5.3.0
PyJWT~=2.8.0
#social-auth-app-django==5.0.0
#social-auth-core==5.0.0

Pillow~=10.3.0
django-storages~=1.14.0
django-ckeditor~=6.7.0
psycopg2-binary~=2.9.0
python-slugify~=8.0.0
requests~=2.32.0
whitenoise~=6.6.0
gunicorn~=22.0.0
boto3~=1.34.0
botocore~=1.34.0
s3transfer~=0.10.0
